requirements.txt
aiohttp==3.7.3
alpaca-trade-api==0.51.0
alpha-vantage==2.3.1
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
multidict==5.1.0
numpy==1.19.4
pandas==1.2.0
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.25.11
websocket-client==0.57.0
websockets==8.1
yarl==1.6.3
